build(refactor): usage instructions for mcp
